Move <Bar> into a custom element
